requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
langchain
langchain-groq
mcp
fastmcp
mcp-use
langchain-mcp-adapters
python-dotenv
langchain-classic
langchain-experimental
langchain-huggingface
faiss-cpu
fastembed
docling
langchain-docling
langchain-milvus
sentence-transformers
pymilvus
pymilvus[milvus_lite]
chromadb
rich
